requirements.txt
Django
whitenoise
mysqlclient
python-dotenv
django-rest-framework